pyproject.toml
[tool.poetry]
name = "wsimorph"
version = "0.1.0"
description = ""
authors = ["Hugh Warden <hugh.warden@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
openslide-python = "^1.4.3"
humanize = "^4.15.0"
pillow = "^12.0.0"
numpy = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "sphinx (>=9.1.0,<10.0.0)",
    "sphinx-rtd-theme (>=3.1.0,<4.0.0)"
]
